Move GP Api entity mapping to a separate file
