+++ GlobalPayments-iOS-SDK/Classes/Mapping/GpApiMapping.swift
@@ -1,0 +1,171 @@
+import Foundation
+
+public struct GpApiMapping {
+
+    public static func mapTransaction(_ doc: JsonDoc?) -> Transaction {
+
+        let transaction = Transaction()
+        transaction.transactionId = doc?.getValue(key: "id")
+        if let amount: String = doc?.getValue(key: "amount") {
+            transaction.balanceAmount = NSDecimalNumber(string: amount).amount
+        }
+        transaction.timestamp = doc?.getValue(key: "time_created")
+        transaction.responseMessage = doc?.getValue(key: "status")
+        transaction.referenceNumber = doc?.getValue(key: "reference")
+        let batchSummary = BatchSummary()
+        batchSummary.sequenceNumber = doc?.getValue(key: "batch_id")
+        transaction.batchSummary = batchSummary
+        transaction.responseCode = doc?.get(valueFor: "action")?.getValue(key: "result_code")
+        transaction.token = doc?.getValue(key: "id")
+
+        if let paymentMethod: JsonDoc = doc?.get(valueFor: "payment_method"),
+           let card: JsonDoc = paymentMethod.get(valueFor: "card") {
+            transaction.cardLast4 = card.getValue(key: "masked_number_last4")
+        }
+
+        if let card: JsonDoc = doc?.get(valueFor: "card") {
+            transaction.cardExpMonth = Int(card.getValue(key: "expiry_month") ?? .empty)
+            transaction.cardExpYear = Int(card.getValue(key: "expiry_year") ?? .empty)
+            transaction.cardNumber = card.getValue(key: "number")
+            transaction.cardType = card.getValue(key: "brand")
+        }
+
+        return transaction
+    }
+
+    public static func mapTransactionSummary(_ doc: JsonDoc?) -> TransactionSummary {
+        let paymentMethod: JsonDoc? = doc?.get(valueFor: "payment_method")
+        let card: JsonDoc? = paymentMethod?.get(valueFor: "card")
+
+        let summary = TransactionSummary()
+        //TODO: Map all transaction properties
+        summary.transactionId = doc?.getValue(key: "id")
+        let timeCreated: String? = doc?.getValue(key: "time_created")
+        summary.transactionDate = timeCreated?.format()
+        summary.transactionStatus = TransactionStatus(value: doc?.getValue(key: "status"))
+        summary.transactionType = doc?.getValue(key: "type")
+        summary.channel = doc?.getValue(key: "channel")
+        summary.amount = NSDecimalNumber(string: doc?.getValue(key: "amount")).amount
+        summary.currency = doc?.getValue(key: "currency")
+        summary.referenceNumber = doc?.getValue(key: "reference")
+        summary.clientTransactionId = doc?.getValue(key: "reference")
+        // ?? = doc.getValue(key: "time_created_reference")
+        summary.batchSequenceNumber = doc?.getValue(key: "batch_id")
+        summary.country = doc?.getValue(key: "country")
+        // ?? = doc.getValue(key: "action_create_id")
+        summary.originalTransactionId = doc?.getValue(key: "parent_resource_id")
+
+        summary.gatewayResponseMessage = paymentMethod?.getValue(key: "message")
+        summary.entryMode = paymentMethod?.getValue(key: "entry_mode")
+        //?? = paymentMethod?.getValue(key: "name")
+
+        summary.cardType = card?.getValue(key: "brand")
+        summary.authCode = card?.getValue(key: "authcode")
+        summary.brandReference = card?.getValue(key: "brand_reference")
+        summary.aquirerReferenceNumber = card?.getValue(key: "arn")
+        summary.maskedCardNumber = card?.getValue(key: "masked_number_first6last4")
+
+        summary.depositReference = doc?.getValue(key: "deposit_id")
+        let depositTimeCreated: String? = doc?.getValue(key: "deposit_time_created")
+        summary.depositDate = depositTimeCreated?.format("YYYY-MM-DD")
+        summary.depositStatus = DepositStatus(value: doc?.getValue(key: "deposit_status"))
+
+        return summary
+    }
+
+    public static func mapDepositSummary(_ doc: JsonDoc?) -> DepositSummary {
+        let summary = DepositSummary()
+        summary.depositId = doc?.getValue(key: "id")
+        let timeCreated: String? = doc?.getValue(key: "time_created")
+        summary.depositDate = timeCreated?.format("yyyy-MM-dd")
+        summary.status = doc?.getValue(key: "status")
+        summary.type = doc?.getValue(key: "funding_type")
+        summary.amount = NSDecimalNumber(string: doc?.getValue(key: "amount")).amount
+        summary.currency = doc?.getValue(key: "currency")
+
+        summary.merchantNumber = doc?.get(valueFor: "system")?.getValue(key: "mid")
+        summary.merchantHierarchy = doc?.get(valueFor: "system")?.getValue(key: "hierarchy")
+        summary.merchantName = doc?.get(valueFor: "system")?.getValue(key: "name")
+        summary.merchantDbaName = doc?.get(valueFor: "system")?.getValue(key: "dba")
+
+        summary.salesTotalCount = doc?.get(valueFor: "sales")?.getValue(key: "count")
+        summary.amount = NSDecimalNumber(string: doc?.get(valueFor: "sales")?.getValue(key: "amount")).amount
+
+        summary.refundsTotalCount = doc?.get(valueFor: "refunds")?.getValue(key: "count")
+        summary.refundsTotalAmount = NSDecimalNumber(string: doc?.get(valueFor: "refunds")?.getValue(key: "amount")).amount
+
+        summary.chargebackTotalCount = doc?.get(valueFor: "disputes")?.get(valueFor: "chargebacks")?.getValue(key: "count")
+        summary.chargebackTotalAmount = NSDecimalNumber(string: doc?.get(valueFor: "disputes")?.get(valueFor: "chargebacks")?.getValue(key: "amount")).amount
+
+        summary.adjustmentTotalCount = doc?.get(valueFor: "disputes")?.get(valueFor: "reversals")?.getValue(key: "count")
+        summary.adjustmentTotalAmount = NSDecimalNumber(string: doc?.get(valueFor: "disputes")?.get(valueFor: "reversals")?.getValue(key: "amount")).amount
+
+        summary.feesTotalAmount = NSDecimalNumber(string: doc?.get(valueFor: "fees")?.getValue(key: "amount")).amount
+
+        return summary
+    }
+
+    public static func mapDisputeSummary(_ doc: JsonDoc?) -> DisputeSummary {
+        let summary = DisputeSummary()
+        summary.caseId = doc?.getValue(key: "id")
+        let timeCreated: String? = doc?.getValue(key: "time_created")
+        summary.caseIdTime = timeCreated?.format()
+        summary.caseStatus = doc?.getValue(key: "status")
+        summary.caseStage = DisputeStage(value: doc?.getValue(key: "stage"))
+        //stage
+        summary.caseAmount = NSDecimalNumber(string: doc?.getValue(key: "amount")).amount
+        summary.caseCurrency = doc?.getValue(key: "currency")
+        summary.caseMerchantId = doc?.get(valueFor: "system")?.getValue(key: "mid")
+        summary.merchantHierarchy = doc?.get(valueFor: "system")?.getValue(key: "hierarchy")
+        summary.transactionMaskedCardNumber = doc?.get(valueFor: "payment_method")?.get(valueFor: "card")?.getValue(key: "number")
+        summary.transactionARN = doc?.get(valueFor: "payment_method")?.get(valueFor: "card")?.getValue(key: "arn")
+        summary.transactionCardType = doc?.get(valueFor: "payment_method")?.get(valueFor: "card")?.getValue(key: "brand")
+        //reason_code
+        summary.reason = doc?.getValue(key: "reason_description")
+        summary.reasonCode = doc?.getValue(key: "reason_code")
+        let timeToRespondBy: String? = doc?.getValue(key: "time_to_respond_by")
+        summary.respondByDate = timeToRespondBy?.format()
+        if let documents: [JsonDoc] = doc?.getValue(key: "documents"), !documents.isEmpty {
+            summary.documents = documents.compactMap {
+                guard let id: String = $0.getValue(key: "id"),
+                      let type = DocumentType(value: $0.getValue(key: "type")) else { return nil }
+                return DisputeDocument(id: id, type: type)
+            }
+        }
+
+        summary.lastAdjustmentFunding = AdjustmentFunding(value: doc?.getValue(key: "last_adjustment_funding"))
+        summary.lastAdjustmentAmount = NSDecimalNumber(string: doc?.getValue(key: "last_adjustment_amount")).amount
+        summary.lastAdjustmentCurrency = doc?.getValue(key: "last_adjustment_currency")
+
+        summary.result = doc?.getValue(key: "result")
+
+        return summary
+    }
+
+    public static func mapDisputeAction(_ doc: JsonDoc?) -> DisputeAction {
+        let action = DisputeAction()
+        action.reference = doc?.getValue(key: "id")
+        action.status = DisputeStatus(value: doc?.getValue(key: "status"))
+        action.stage = DisputeStage(value: doc?.getValue(key: "stage"))
+        action.amount = NSDecimalNumber(string: doc?.getValue(key: "amount")).amount
+        action.currency = doc?.getValue(key: "currency")
+        action.reasonCode = doc?.getValue(key: "reason_code")
+        action.reasonDescription = doc?.getValue(key: "reason_description")
+        action.result = DisputeResult(value: doc?.getValue(key: "result"))
+        if let documents: [JsonDoc?] = doc?.getValue(key: "documents") {
+            action.documents = documents.compactMap { $0?.getValue(key: "id") }
+        }
+
+        return action
+    }
+
+    public static func mapDocumentMetadata(_ doc: JsonDoc?) -> DocumentMetadata? {
+        guard let id: String = doc?.getValue(key: "id"),
+              let b64Content: String = doc?.getValue(key: "b64_content"),
+              let convertedData = Data(base64Encoded: b64Content) else {
+            return nil
+        }
+
+        return DocumentMetadata(id: id, b64Content: convertedData)
+    }
+}
